[develop]: DB 스키마 및 Entity 정리

+++ src/main/java/app/hdj/datepick/domain/review/dto/ModifyPlaceReviewDto.java
@@ -10,7 +10,7 @@ import lombok.Getter;
 public class ModifyPlaceReviewDto {
     //place Order 만 지정하면 place Id 는 알아서 지정해주도록.
     private Byte placeOrder;
-    private Byte rating;
+    private Float rating;
     private String content;
 }
